Working prototype with tests.
When the plugin detects a patchset has been created, it will
merge downstream until it hits a merge conflict. On the
conflicting merge, it will vote -1 on a configurable label
and provide instructions to resolving the merge conflict.

Draft changes will be ignored until published.

The plugin will put all the auto-created changes in the
same topic as the original change (or create a topic if
none exists). If a user updates the topic, it will update
the topic of all the downstream merges.

If there are existing downstream merges from a previous
automerged patchset, it will abandon them all without
email spamming the end user.

Admins can update the config.yaml file in tools/automerger
(this is configurable) in order to add merge paths and
do other configuration changes.

Javascript portion will bold branches with the same
branch in the RelatedChanges sections (including
Same Topic), and add a button to skip all downstream
changes.

Change-Id: I89209ed29073f987a37d03e10dad85a1c5c530fb