Update patch set 2
Patch Set 2:

I'm having trouble tracking what DownstreamCreator does. Could there be a comment (either in the code or the commit message) describing what it wants to do?

In particular I keep getting confused about how it decides what revision to work with --- the revision that the user voted on, or the current revision for the change.

Patch-set: 2
Label: Verified=0