Update patch set 5
Patch Set 5:

(6 comments)

Sorry for the slow turnaround.

I found this hard to review because it is doing so many things at once:
* adding docs
* reordering code
* some fixes, like changing log.info to log.debug for an expected condition
* adding the on-change-comment listener

I get lost when switching between things from file to file, so a single change per intended effect tends to be easier to review.

Patch-set: 5
Label: Verified=0