Update patch set 4
Patch Set 4:

I was originally trying to be as unobtrusive as possible (i.e. failure on plugin should be silent to avoid annoying people who aren't using the plugin) but in retrospect that's not a good strategy. If something broken there we should throw an error to let people know about it instead of failing silently.

Patch-set: 4