Update patch set 15
Patch Set 15:

Friendly ping? :D

This is probably the highest priority item for me in terms of being able to ramp dogfood up to more teams, as it allows me to change the config and have it take effect immediately (and thus be able to turn this on and off for teams by just updating the config file).

Patch-set: 15
Reviewer: Gerrit User 1019952 <1019952@173816e5-2b9a-37c3-8a2e-48639d4f1153>