Allow configurable vote values.
Since the label is configurable, the vote values should be
as well. Users of the plugin may want to be able to vote
Code-Review -2 on a merge conflict, or Verified +1 on a
successful merge.

We also now default to Code-Review instead of Verified
for voting, since Verified may not be installed by
default.

Change-Id: I8aed1d313641c47f2e15f7acad083b2fab7e1c76