Support multi-word topics or topics with quotes/braces.
Currently, if there is a multi-word topic and someone tries
to submit, we error out with the missingDownstreamsMessage.
This is because we have a badly formatted query when we try
to see if there are any missing downstream branches in the
merge validator.

Instead, we should surround the topic in quotes or braces
so that we don't treat it as a separate query argument. We
do this by using the newly created QueryBuilder class to
construct our queries with escaped parameters for us.

Change-Id: I259df3cdabdc0b5e4f39794b5f976438d8e2a7ec