Support multi-word topics and topics with quotes or braces (but not both).
Currently, if there is a multi-word topic and someone tries
to submit, we error out with the missingDownstreamsMessage.
This is because we have a badly formatted query when we try
to see if there are any missing downstream branches in the
merge validator.

Instead, we should surround the topic in quotes or braces so that we
don't treat it as a separate query argument.

Note: Gerrit seems to fail to submit together if you try to submit a topic that has quotes, so changed the default missingDownstreamsMessage to inform of that case.

Change-Id: I259df3cdabdc0b5e4f39794b5f976438d8e2a7ec