Support multi-word topics.
Currently, if there is a multi-word topic and someone tries
to submit, we error out with the missingDownstreamsMessage.
This is because we have a badly formatted query when we try
to see if there are any missing downstream branches in the
merge validator.

Instead, we should surround the topic in quotes so that we
don't treat it as a separate query argument.

Change-Id: I259df3cdabdc0b5e4f39794b5f976438d8e2a7ec