Handle diamond merges correctly.
If we have branches top, left, right, and bottom, and left == right,
then creating a change on top should create two changes with the same
sha on left and right, and only a single change on bottom.

top -> left -> bottom, and top -> right -> bottom simultaneously.

DO NOT SUBMIT YET: test is flaky, since when we upload a change to top,
left and right changes MIGHT or MIGHT NOT have the same sha based on
whether or not they are created in the same second. Need to find a way
to make sure that they have the same sha in the test.

Change-Id: I8813cefb7cb7052b435dbef4e1af422c516e13a0