Stacked changes continue being stacked downstream.
Previously, if change A and B were uplaoded, where Bs first parent was
A, then the first parents of both A and B were HEAD of the downstream
branch.
Now, if changes A and B are uploaded, where Bs first parent is A, then for
downstream A' and B', the first parent of B' should be A'.

Change-Id: I09a9abc6320129f64958859efc2e379214903e54