Use credentials of configurable contextUserId for API calls.
Allow setting contextUserId in the config so that all automerger created
CLs and votes can run using the permissions of that user. This will
allow +2s and other things that are normally restricted for a user to
work downstream under the credentials of a robot account.

Not setting contextUserId will mean that a user's credentials will
continue to be the credentials used downstream.

Change-Id: If861a2cffdb68911fb7821c568a3fef67c0d0ddc