Remove DownstreamCreatorIT#testDiamondMerge_identicalDiamondSides
This test froze the time to create 2 changes with the same commit SHA1.
Having 2 changes with the same commit SHA1 (on the same target branch)
is an inconsistency that Gerrit cannot deal well with. This is why
Gerrit was fixed so that changes which are created with the same
content, on the same base commit and at the same timestamp have
different SHA1s now. This is achieved by generating random Change-Ids,
instead of computing them (see change Ie4f01e5a3). Since the Change-Id
is part of the commit message, each change has a different SHA1 now.
Since the test tested a situation with inconsitent data in Gerrit which
can no longer occur, the test should be dropped.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If10a1ca63f4dc54537dca692132ed9ccfef21f83