Perform work asynchronously
The automerger plugin performs expensive work on the request path. The
work isn't actually relevant for the request because it is not returned to the client. To the best of our knowledge, there are no clients that
immediately read back the actions triggered on downstream changes.

To meet SLOs, we do this work asynchrounously. Since the event data
types are mutable, we take a pragmatic approach to make this thread
safe that is deep copying the objects using GSON.

A failure in the old synchronous path had a similar effect as in the
newly asynchronous path (i.e. the downstream work is dropped).

Release-Notes: skip
Change-Id: I6693e014acc682f01bbf03af9219f1bb6969a131