Remove now-unused import of `Truth8.assertThat`.
Since Truth
[1.3.0](https://github.com/google/truth/releases/tag/v1.3.0), the main
`Truth` class has an `assertThat` overload for `Optional`. Since this
class already imports `Truth.assertThat`, it uses that for its call to
`assertThat(getCodeReview(id)).isEmpty()`, so it does not need the
import of `Truth8.assertThat`.

In fact, importing both `Truth8.assertThat` and `Truth.assertThat` will
produce a "reference to assertThat is ambiguous" error in an upcoming
version of Truth.

Change-Id: I80413ed955d871dd3f9e744eca4f467e7c988fac